pyproject.toml
[project]
name = "astrolabe"
version = "0.1.0"
description = "TLE Tracker"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "supabase>=2.0.0",
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
    "paramiko>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
populate-tles = "scripts.populate_tles:main"
update-tles = "scripts.update_tles:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ipython>=8.0.0",
]